+++ hutool-crypto/src/test/java/cn/hutool/crypto/test/ECIESTest.java
@@ -1,0 +1,26 @@
+package cn.hutool.crypto.test;
+
+import cn.hutool.core.util.StrUtil;
+import cn.hutool.crypto.asymmetric.ECIES;
+import cn.hutool.crypto.asymmetric.KeyType;
+import org.junit.Assert;
+import org.junit.Test;
+
+public class ECIESTest {
+
+	@Test
+	public void eciesTest(){
+		final ECIES ecies = new ECIES();
+
+		String textBase = "我是一段特别长的测试";
+		StringBuilder text = new StringBuilder();
+		for (int i = 0; i < 10; i++) {
+			text.append(textBase);
+		}
+
+		// 公钥加密，私钥解密
+		String encryptStr = ecies.encryptBase64(text.toString(), KeyType.PublicKey);
+		String decryptStr = StrUtil.utf8Str(ecies.decrypt(encryptStr, KeyType.PrivateKey));
+		Assert.assertEquals(text.toString(), decryptStr);
+	}
+}